Seperate fuzzy function
